tests/wpt/tests/css/css-text/parsing/text-wrap-pretty.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: text-wrap: pretty parsing</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-wrap">
<meta name="assert" content="text-wrap: pretty parsing">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_valid_value("text-wrap", "pretty");
test_valid_value("text-wrap", "wrap pretty", "pretty");
test_valid_value("text-wrap", "pretty wrap", "pretty");
test_valid_value("text-wrap", "stable wrap", "stable");
test_valid_value("text-wrap", "nowrap pretty");
test_valid_value("text-wrap", "pretty nowrap", "nowrap pretty");
test_valid_value("text-wrap-style", "pretty");

test_computed_value("text-wrap", "pretty");
test_computed_value("text-wrap", "wrap pretty", "pretty");
test_computed_value("text-wrap", "pretty wrap", "pretty");
test_computed_value("text-wrap", "stable wrap", "stable");
test_computed_value("text-wrap", "nowrap pretty");
test_computed_value("text-wrap", "pretty nowrap", "nowrap pretty");
test_computed_value("text-wrap-style", "pretty");
</script>
</body>
</html>
