third_party/blink/web_tests/external/wpt/css/css-text/parsing/text-wrap-pretty-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['text-wrap'] = "wrap pretty" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap'] = "pretty wrap" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap'] = "stable wrap" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap'] = "nowrap pretty" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap'] = "pretty nowrap" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['text-wrap-style'] = "pretty" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] Property text-wrap value 'wrap pretty'
  assert_true: 'wrap pretty' is a supported value for text-wrap. expected true got false
[FAIL] Property text-wrap value 'pretty wrap'
  assert_true: 'pretty wrap' is a supported value for text-wrap. expected true got false
[FAIL] Property text-wrap value 'stable wrap'
  assert_true: 'stable wrap' is a supported value for text-wrap. expected true got false
[FAIL] Property text-wrap value 'nowrap pretty'
  assert_true: 'nowrap pretty' is a supported value for text-wrap. expected true got false
[FAIL] Property text-wrap value 'pretty nowrap'
  assert_true: 'pretty nowrap' is a supported value for text-wrap. expected true got false
[FAIL] Property text-wrap-style value 'pretty'
  assert_true: text-wrap-style doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
